% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot_helpers.R
\name{add_point.PearsonDiagram}
\alias{add_point.PearsonDiagram}
\title{Add a point to the Pearson Diagram}
\usage{
\method{add_point}{PearsonDiagram}(object, sq_skewness, kurtosis, distribution = NA)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{sq_skewness}{Square of skewness value of the distribution}

\item{kurtosis}{Kurtosis value of the distribution}

\item{distribution}{The name of the distribution}
}
\value{
Updated PearsonDiagram object with new points
}
\description{
Add a point to the Pearson Diagram
}
