% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{add_point}
\alias{add_point}
\title{Add a point to the Pearson Diagram}
\usage{
add_point(object, sq_skewness, kurtosis, distribution = NA)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{sq_skewness}{Square of skewness value of the distribution}

\item{kurtosis}{Kurtosis value of the distribution}

\item{distribution}{The name of the distribution}
}
\value{
Updated PearsonDiagram object with new points
}
\description{
Adds a point with the given sq_skewness and kurtosis values.
}
