% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{add_points_from_list}
\alias{add_points_from_list}
\title{Add points to the Pearson Diagram from a list of datasets}
\usage{
add_points_from_list(object, input_list)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{input_list}{A list of numeric vectors representing datasets}
}
\value{
Updated PearsonDiagram object with new points
}
\description{
This function adds multiple points to the Pearson Diagram, calculated from the given list of datasets.
}
