% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{add_unknown_distribution}
\alias{add_unknown_distribution}
\title{Add Unknown Distribution to Pearson Diagram}
\usage{
add_unknown_distribution(object, data)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{data}{A numeric vector representing the unknown distribution}
}
\value{
Updated PearsonDiagram object with the unknown distribution point added
}
\description{
This function calculates the skewness and kurtosis for an unknown distribution and adds it to the Pearson diagram.
The unknown distribution is plotted as a black "X" on the diagram.
}
