% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_samples.list}
\alias{bootstrap_samples.list}
\title{Bootstrap Samples for List Input}
\usage{
\method{bootstrap_samples}{list}(object, n_samples = 100, sample_size = NULL)
}
\arguments{
\item{object}{A list of numeric vectors}

\item{n_samples}{The number of bootstrap samples to generate}

\item{sample_size}{The size of each bootstrap sample}
}
\value{
A list of bootstrap samples with calculated moments
}
\description{
This function generates bootstrap samples for a list of numeric vectors.
}
