% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_samples.numeric}
\alias{bootstrap_samples.numeric}
\title{Bootstrap Samples for Numeric Input}
\usage{
\method{bootstrap_samples}{numeric}(object, n_samples = 100, sample_size = NULL)
}
\arguments{
\item{object}{A numeric vector}

\item{n_samples}{The number of bootstrap samples to generate}

\item{sample_size}{The size of each bootstrap sample}
}
\value{
A list of bootstrap samples with calculated moments
}
\description{
This function generates bootstrap samples for numeric vectors directly.
}
