% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_samples.PearsonDiagram}
\alias{bootstrap_samples.PearsonDiagram}
\title{Bootstrap Samples for PearsonDiagram}
\usage{
\method{bootstrap_samples}{PearsonDiagram}(object, n_samples = 100, sample_size = NULL)
}
\arguments{
\item{object}{A PearsonDiagram object}

\item{n_samples}{The number of bootstrap samples to generate}

\item{sample_size}{The size of each bootstrap sample}
}
\value{
A list of bootstrap samples with calculated moments
}
\description{
Bootstrap Samples for PearsonDiagram
}
