% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Generate Bootstrap Samples}
\usage{
bootstrap_samples(input_data, n_samples = 100, sample_size = NULL)
}
\arguments{
\item{input_data}{A numeric vector or list of numeric vectors}

\item{n_samples}{The number of bootstrap samples to generate}

\item{sample_size}{The size of each bootstrap sample (default: same as the length of the input data)}
}
\value{
A list of bootstrap samples with calculated moments
}
\description{
This function generates bootstrap samples from the provided input data.
}
