% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{calculate_bootstrap_moments}
\alias{calculate_bootstrap_moments}
\title{Calculate Moments for Bootstrap Samples}
\usage{
calculate_bootstrap_moments(bootstrap_samples)
}
\arguments{
\item{bootstrap_samples}{A list of numeric vectors representing bootstrap samples}
}
\value{
A data frame containing skewness and kurtosis for each sample
}
\description{
This function calculates skewness and kurtosis for each bootstrap sample.
}
