% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{canvas_creation}
\alias{canvas_creation}
\title{Create Pearson Diagram Canvas}
\usage{
canvas_creation()
}
\value{
A ggplot2 object representing the Pearson diagram canvas.
}
\description{
This function creates a Pearson diagram canvas with known distributions without any input data points.
The canvas represents Normal, Exponential, Gamma, and Beta distributions.
}
