% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{compare_distributions}
\alias{compare_distributions}
\title{Compare multiple distributions}
\usage{
compare_distributions(object, data_list)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{data_list}{A list of numeric vectors}
}
\value{
Updated PearsonDiagram object with new points and a plot
}
\description{
Compare multiple distributions
}
