% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{compare_multiple_distributions.PearsonDiagram}
\alias{compare_multiple_distributions.PearsonDiagram}
\title{Compare Multiple Distributions on Pearson Diagram}
\usage{
compare_multiple_distributions.PearsonDiagram(object, data_list)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{data_list}{A list of numeric vectors}
}
\value{
Updated PearsonDiagram object with new points and a plot
}
\description{
This function compares multiple distributions, validates the data, and adds them to the Pearson Diagram.
}
