% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{compare_multiple_distributions}
\alias{compare_multiple_distributions}
\title{Compare and Plot Multiple Distributions on Pearson Diagram}
\usage{
compare_multiple_distributions(object, n_samples = 100, sample_size = 1000)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{n_samples}{Number of samples to generate for each distribution}

\item{sample_size}{Size of each sample}
}
\value{
Updated PearsonDiagram object with new points and a plot
}
\description{
This function generates samples from various distributions (Normal, Exponential, Uniform, Beta, Gamma),
calculates their skewness and kurtosis, and plots them on the Pearson Diagram.
}
