% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{cpp_calculate_moments}
\alias{cpp_calculate_moments}
\title{Calculate Moments Using Rcpp}
\usage{
cpp_calculate_moments(data)
}
\arguments{
\item{data}{A numeric vector containing the data points}
}
\value{
A numeric vector with two values: square of skewness and kurtosis
}
\description{
This function calculates the skewness and kurtosis using the Rcpp implementation.
}
