% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{cpp_calculate_moments_for_list}
\alias{cpp_calculate_moments_for_list}
\title{Calculate Moments Using Rcpp for a List of Vectors}
\usage{
cpp_calculate_moments_for_list(data_list)
}
\arguments{
\item{data_list}{A list of numeric vectors containing data points.}
}
\value{
A list containing the skewness and kurtosis for each vector.
}
\description{
This function calculates skewness and kurtosis using Rcpp for a list of numeric vectors.
}
