% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{create_pearson_canvas}
\alias{create_pearson_canvas}
\title{Create Pearson Diagram Canvas}
\usage{
create_pearson_canvas()
}
\value{
A ggplot2 object for the Pearson diagram.
}
\description{
Creates the base Pearson diagram with known distributions.
}
