% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{ensure_python_dependencies}
\alias{ensure_python_dependencies}
\title{Ensure required Python dependencies are installed}
\usage{
ensure_python_dependencies()
}
\description{
This function ensures that the necessary Python packages for rendering plots
are installed (kaleido and plotly).
It uses reticulate to install the packages via conda.
}
