% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{generate_area_data}
\alias{generate_area_data}
\title{Generate Data for Known Distributions - Area Representations}
\usage{
generate_area_data()
}
\value{
A data frame containing skewness, kurtosis, and parameter space for multiple distributions.
}
\description{
This function generates skewness and kurtosis values for known distributions representing area in pearson diagram using the calculate_moments function.
}
