% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{generate_data}
\alias{generate_data}
\title{Generate Data for Known Distributions}
\usage{
generate_data()
}
\value{
A data frame containing skewness, kurtosis, and distribution names for multiple distributions.
}
\description{
This function generates skewness and kurtosis values for various known distributions using the calculate_moments function.
}
