% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot_helpers.R
\name{highlight_point_on_hover}
\alias{highlight_point_on_hover}
\title{Highlight Points on Pearson Diagram with Hover Information}
\usage{
highlight_point_on_hover(p)
}
\arguments{
\item{p}{A ggplot2 object with points plotted on it.}
}
\value{
A plotly object with added interactivity.
}
\description{
This function adds interactive tooltip functionality to Pearson diagram plots using plotly.
}
