% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot_helpers.R
\name{PearsonDiagram}
\alias{PearsonDiagram}
\title{Create a PearsonDiagram Object}
\usage{
PearsonDiagram()
}
\value{
A PearsonDiagram object.
}
\description{
This function creates an empty PearsonDiagram object.
The PearsonDiagram object is used to store sq_skewness, kurtosis values, and distribution.
}
\examples{
pd <- PearsonDiagram()
}
