% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{plot_diagram.PearsonDiagram}
\alias{plot_diagram.PearsonDiagram}
\title{Plot the Pearson Diagram}
\usage{
\method{plot_diagram}{PearsonDiagram}(object)
}
\arguments{
\item{object}{PearsonDiagram object}
}
\value{
ggplot2 object representing the Pearson diagram
}
\description{
Plot the Pearson Diagram
}
