% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{plot_diagram_with_unknown}
\alias{plot_diagram_with_unknown}
\title{Plot the Pearson Diagram with Unknown Distribution}
\usage{
plot_diagram_with_unknown(object)
}
\arguments{
\item{object}{PearsonDiagram object}
}
\value{
ggplot2 object representing the Pearson diagram with an unknown distribution
}
\description{
Plot the Pearson Diagram with Unknown Distribution
}
