% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{plot_pearson_diagram}
\alias{plot_pearson_diagram}
\title{Plot Pearson Diagram}
\usage{
plot_pearson_diagram(skewness, kurtosis)
}
\arguments{
\item{skewness}{The skewness of the data}

\item{kurtosis}{The kurtosis of the data}
}
\value{
A ggplot2 object representing the Pearson diagram
}
\description{
This function plots the Pearson diagram based on provided skewness and kurtosis values.
}
\examples{
plot_pearson_diagram(0, 3) # Plot a normal distribution
}
