% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Statistics for Numeric Inputs}
\usage{
summary_stats(x, censored)
}
\arguments{
\item{x}{A numeric vector or a list of numeric vectors for which the summary statistics are to be calculated.}

\item{censored}{binary value indicating if the outliers are removed from the input}
}
\value{
A data frame containing the summary statistics. If the input is a vector, the row is labeled as "sample1".
        If the input is a list, each row corresponds to a sample (e.g., "sample1", "sample2", etc.), and the summary statistics are computed for each vector.
}
\description{
This function calculates summary statistics for numeric inputs, which can either be a numeric vector or a list of numeric vectors.
The summary statistics include the number of observations, number of missing values, mean, median, and standard deviation.
}
