% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{validate_input.list}
\alias{validate_input.list}
\title{Validate List of Numeric Vectors}
\usage{
\method{validate_input}{list}(object)
}
\arguments{
\item{object}{A list of numeric vectors}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
This function checks if the input is a valid list of numeric vectors, ensuring each element meets validation criteria.
}
