% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{validate_input.numeric}
\alias{validate_input.numeric}
\title{Validate Numeric Vector}
\usage{
\method{validate_input}{numeric}(object)
}
\arguments{
\item{object}{A numeric vector}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
This function checks if the input is a valid numeric vector with no missing values or zero variance.
}
