% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{validate_input.PearsonDiagram}
\alias{validate_input.PearsonDiagram}
\title{Validate PearsonDiagram Object}
\usage{
\method{validate_input}{PearsonDiagram}(object)
}
\arguments{
\item{object}{A PearsonDiagram object}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
This function validates if a PearsonDiagram object has the correct structure.
}
