% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{add_point}
\alias{add_point}
\title{Add a point to the Pearson Diagram}
\usage{
add_point(object, skewness, kurtosis)
}
\arguments{
\item{object}{PearsonDiagram object}

\item{skewness}{Skewness value of the distribution}

\item{kurtosis}{Kurtosis value of the distribution}
}
\value{
Updated PearsonDiagram object with new points
}
\description{
Add a point to the Pearson Diagram
}
