% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Bootstrap Samples for PearsonDiagram (S3 Method)}
\usage{
bootstrap_samples(object, n_samples = 100, sample_size = NULL)
}
\arguments{
\item{object}{A PearsonDiagram object}

\item{n_samples}{The number of bootstrap samples to generate}

\item{sample_size}{The size of each bootstrap sample (default: same as the length of the input data)}
}
\value{
A list of bootstrap samples with calculated moments
}
\description{
This function generates bootstrap samples from the provided input data for a PearsonDiagram object.
}
