% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{highlight_point_on_hover}
\alias{highlight_point_on_hover}
\title{Highlight Points on Pearson Diagram}
\usage{
highlight_point_on_hover(point)
}
\arguments{
\item{point}{A ggplot2 object with points plotted on it}
}
\value{
A modified ggplot2 object
}
\description{
This function adds interactive tooltip functionality to Pearson diagram plots.
}
