% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{PearsonDiagram}
\alias{PearsonDiagram}
\title{Create a PearsonDiagram Object}
\usage{
PearsonDiagram()
}
\value{
An empty PearsonDiagram object.
}
\description{
This function creates an empty PearsonDiagram object. The PearsonDiagram object is used to store skewness and kurtosis values and allows for plotting and comparing distributions.
}
\examples{
pd <- PearsonDiagram()
}
