% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{plot_diagram}
\alias{plot_diagram}
\title{Plot the Pearson Diagram with Input Data and Bootstrap Samples}
\usage{
plot_diagram(
  object,
  input_data = NULL,
  bootstrap = FALSE,
  hover = TRUE,
  treat.outliers = FALSE,
  summary.file = NULL,
  plot.name = NULL,
  title.font.family = "Arial",
  title.font.size = 16,
  title.font.color = "black",
  title.hjust = NULL,
  legend.font.family = "Arial",
  legend.font.size = 12,
  legend.font.color = "black",
  legend.hjust = NULL,
  axis.font.family = "Arial",
  axis.font.size = 14,
  axis.font.color = "black"
)
}
\arguments{
\item{object}{A PearsonDiagram object.}

\item{input_data}{A numeric vector or a list of numeric vectors containing the data points (optional).}

\item{bootstrap}{A boolean indicating whether to perform bootstrap analysis.}

\item{hover}{A boolean indicating whether to add hover functionality.}

\item{treat.outliers}{A boolean indicating whether to exclude extreme outliers (default is FALSE).}

\item{summary.file}{Filename of the summary report}

\item{plot.name}{Filename of the output plot}

\item{title.font.family}{Font family of the plot title}

\item{title.font.size}{Font size of the plot title}

\item{title.font.color}{Font color of the plot title}

\item{title.hjust}{alignment of the plot title}

\item{legend.font.family}{Font family of the plot legend}

\item{legend.font.size}{Font size of the plot legend}

\item{legend.font.color}{Font color of the plot legend}

\item{legend.hjust}{alignment of the plot legend}

\item{axis.font.family}{Font family of the plot axis}

\item{axis.font.size}{Font size of the plot axis}

\item{axis.font.color}{Font color of the plot axis}
}
\value{
A ggplot2 object representing the Pearson diagram.
}
\description{
This function plots the Pearson diagram with known distributions and user-provided data points.
It supports input_data as a numeric vector or a list of numeric vectors. For each input, it calculates the
skewness and kurtosis and plots them on the canvas.
}
