% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_plot.R
\name{plot_diagram}
\alias{plot_diagram}
\title{Plot the Pearson Diagram}
\usage{
plot_diagram(object)
}
\arguments{
\item{object}{PearsonDiagram object}
}
\value{
ggplot2 object representing the Pearson diagram
}
\description{
Plot the Pearson Diagram
}
