% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_diagram}
\alias{plot_diagram}
\title{Plot the Pearson Diagram}
\usage{
plot_diagram(
  input_data = NULL,
  bootstrap = FALSE,
  treat.outliers = FALSE,
  summary.file = NULL,
  plot.name = NULL,
  title.font.family = "Arial",
  title.font.size = 16,
  title.font.color = "black",
  title.hjust = NULL,
  legend.font.family = "Arial",
  legend.font.size = 12,
  legend.font.color = "black",
  legend.hjust = NULL,
  axis.font.family = "Arial",
  axis.font.size = 14,
  axis.font.color = "black",
  cb.friendly = FALSE
)
}
\arguments{
\item{input_data}{A numeric vector or a list of numeric vectors containing the data points (optional).}

\item{bootstrap}{A boolean indicating whether to perform bootstrap analysis.}

\item{treat.outliers}{A boolean indicating whether to exclude extreme outliers (default is FALSE).}

\item{summary.file}{Filename of the summary report}

\item{plot.name}{Filename of the output plot}

\item{title.font.family}{Font family of the plot title}

\item{title.font.size}{Font size of the plot title}

\item{title.font.color}{Font color of the plot title}

\item{title.hjust}{alignment of the plot title}

\item{legend.font.family}{Font family of the plot legend}

\item{legend.font.size}{Font size of the plot legend}

\item{legend.font.color}{Font color of the plot legend}

\item{legend.hjust}{alignment of the plot legend}

\item{axis.font.family}{Font family of the plot axis}

\item{axis.font.size}{Font size of the plot axis}

\item{axis.font.color}{Font color of the plot axis}

\item{cb.friendly}{Logical, if TRUE uses a colorblind-friendly palette}
}
\value{
A ggplot2 object representing the Pearson diagram.
}
\description{
This function plots the Pearson diagram with known distributions and user-provided data points.
It supports input_data as a numeric vector or a list of numeric vectors. For each input, it calculates the
skewness and kurtosis and plots them on the canvas.
}
\examples{
# Example 1: Basic plot with a numeric vector
data_vector <- c(1.2, 2.3, 3.4, 4.5, 5.6)
plot_diagram(input_data = data_vector)

# Example 2: Plot with a list of numeric vectors
data_list <- list(c(1.1, 2.2, 3.3), c(2.2, 3.3, 4.4))
plot_diagram(input_data = data_list)

# Example 3: Plot with bootstrap analysis
plot_diagram(input_data = data_vector, bootstrap = TRUE)

# Example 4: Plot with outlier treatment
noisy_data <- c(1, 2, 3, 100, 5, 6)
plot_diagram(input_data = noisy_data, treat.outliers = TRUE, bootstrap = TRUE)

# Example 5: Customizing fonts and plot appearance
plot_diagram(input_data = data_vector,
             title.font.family = "Arial", title.font.size = 20, title.font.color = "darkred",
             legend.font.size = 14, axis.font.size = 12)
}
