% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_helpers.R
\name{validate_input}
\alias{validate_input}
\title{Validate Input (S3 Generic)}
\usage{
validate_input(object)
}
\arguments{
\item{object}{The object to be validated, which could be a PearsonDiagram object, numeric vector, or list.}
}
\value{
TRUE if the input is valid, otherwise an error will be thrown.
}
\description{
This function serves as a generic method for validating inputs.
}
